Organize, fix install_requires
