+++ .gitignore
@@ -1,4 +1,5 @@
 .mypy_cache/
+.pytest_cache
 __pycache__
 *.egg-info
 dist/*
